+++ src/graphql.ts
@@ -1,6 +1,7 @@
 import { octokit } from './octokit'
 import { context } from '@actions/github'
 import { CommittersDetails } from './interfaces'
+import { getPullRequestNumber } from './shared/getPullRequestNumber'
 
 
 
@@ -48,7 +49,7 @@ export default async function getCommitters(): Promise<CommittersDetails[]> {
     }`.replace(/ /g, ''), {
             owner: context.repo.owner,
             name: context.repo.repo,
-            number: context.issue.number,
+            number: getPullRequestNumber(),
             cursor: ''
         })
         response.repository.pullRequest.commits.edges.forEach(edge => {
@@ -56,7 +57,7 @@ export default async function getCommitters(): Promise<CommittersDetails[]> {
             let user = {
                 name: committer.login || committer.name,
                 id: committer.databaseId || '',
-                pullRequestNo: context.issue.number
+                pullRequestNo: getPullRequestNumber()
             }
             if (committers.length === 0 || committers.map((c) => {
                 return c.name